+++ tests/testthat/test-check_impute2_results_unchanged.R
@@ -38,12 +38,20 @@ impute2CoxSurv(impute.file=impute.file,
               clusterObj=NULL,
               keepGDS=FALSE)  
 
-results_file <- read.table(paste0(results_file_name, ".coxph"), sep="\t", header=TRUE, stringsAsFactors = FALSE)
+if(file.exists(paste0(results_file_name, ".coxph"))){
+  results_file <- read.table(paste0(results_file_name, ".coxph"), sep="\t", header=TRUE, stringsAsFactors = FALSE)
+} else {
+  results_file <- NULL
+}
 
 test_that("check impute2 results not changed from example", {
-  skip_on_cran()
-  expect_snapshot_output(results_file)
+  expect_snapshot_output(results_file, cran = F)
 })
 
-file.remove(paste0(results_file_name, ".coxph"))
-file.remove(paste0(results_file_name, ".snps_removed"))
+if(file.exists(paste0(results_file_name, ".coxph"))){
+  file.remove(paste0(results_file_name, ".coxph"))
+}
+
+if(file.exists(paste0(results_file_name, ".snps_removed"))){
+  file.remove(paste0(results_file_name, ".snps_removed"))
+}
